+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Components to display sell data
 
 ### Fixed
 - Change property type of `token` in `Sell` model to `string`